pjsua2: media: Implement AudioMediaPlayer::invalidatePlayer call
This allows us to work around an impedance mismatch with SWIG,
which generates a destructor in its subclass. The AudioMediaPlayer
destructor is responsible for ensuring that callbacks are no
longer invoked after teardown has begun, but SWIG's destructor
runs earlier and removes the mapping for the instance. That leaves
a small window where the onEof() callback may be invoked, causing
a crash.

This change exposes a method that may be called from Go in order
to invoke that destructor logic prior to SWIG removing its mapping,
thus eliminating the possible race.